backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-socketio>=5.9.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.93.0
anthropic>=0.56.0
httpx>=0.25.2
aiofiles>=23.2.1
typing-extensions>=4.11.0